relocate packages
