+++ visual-espresso/visual/build.gradle
@@ -77,7 +77,7 @@ publishing {
 
 apollo {
     service("service") {
-        packageName.set("com.saucelabs.visual.espresso")
+        packageName.set("com.saucelabs.visual.graphql")
         generateKotlinModels.set(false)
         warnOnDeprecatedUsages.set(false)
     }
